backend/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-extensions==3.2.3
django-ninja-extra==0.20.4
django-ninja-jwt==5.3.0
django-ninja==1.1.0
Django==4.2.11
ipdb==0.13.13
pydantic-core==2.16.3
pydantic==2.6.4
sqlparse==0.4.4
typing-extensions==4.10.0